base.js export trick
